users endpoint is auth

+++ courseswebsite/src/main/java/com/medhat/springboot/courseswebsite/rest/UsersRestController.java
@@ -8,6 +8,7 @@ import com.medhat.springboot.courseswebsite.exception.NotAuthorizedException;
 import com.medhat.springboot.courseswebsite.exception.NotFoundException;
 import com.medhat.springboot.courseswebsite.securingweb.WebSecurityConfig;
 import com.medhat.springboot.courseswebsite.securingweb.WebSecurityPermissions;
+import com.medhat.springboot.courseswebsite.service.RolesService;
 import com.medhat.springboot.courseswebsite.service.UsersService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
@@ -23,10 +24,12 @@ import java.util.List;
 public class UsersRestController {
 
     private UsersService usersService;
+    private RolesService rolesService;
 
     @Autowired
-    public UsersRestController(UsersService usersService) {
+    public UsersRestController(UsersService usersService,RolesService rolesService) {
         this.usersService = usersService;
+        this.rolesService = rolesService;
     }
 
     @GetMapping("/users")
@@ -37,7 +40,7 @@ public class UsersRestController {
     @GetMapping("/users/{userId}")
     public Users getById(@PathVariable int userId, Principal principal) {
 
-        if(WebSecurityPermissions.hasPermission(principal,userId,usersService.getById(userId).getUserName()))
+        if(WebSecurityPermissions.hasPermission(principal,usersService.getById(userId).getUserName(),"ADMIN"))
             return usersService.getById(userId);
         else{
             throw new NotAuthorizedException("Access Denied, you don't have permissions to access other users data");
@@ -46,39 +49,71 @@ public class UsersRestController {
 
 
     @GetMapping("/users/{userId}/mycourses")
-    public List<Course> getInstructorCourses(@PathVariable int userId){
-        return usersService.getInstructorCourses(userId);
+    public List<Course> getInstructorCourses(@PathVariable int userId,Principal principal){
+        if(WebSecurityPermissions.hasPermission(principal,usersService.getById(userId).getUserName(),"ADMIN"))
+            return usersService.getInstructorCourses(userId);
+        else
+            throw new NotAuthorizedException("Access Denied, you don't have permissions to access other users data");
     }
 
     @GetMapping("/users/{userId}/enrollments")
-    public List<StudentCoursesData> getEnrolledCourses(@PathVariable int userId){
-        return usersService.getEnrolledCourses(userId);
+    public List<StudentCoursesData> getEnrolledCourses(@PathVariable int userId,Principal principal){
+        if(WebSecurityPermissions.hasPermission(principal,usersService.getById(userId).getUserName(),"ADMIN"))
+            return usersService.getEnrolledCourses(userId);
+        else
+            throw new NotAuthorizedException("Access Denied, you don't have permissions to access other users data");
     }
 
     @PostMapping("/users")
     public Users addUser(@RequestBody Users users){
+
         return usersService.saveUser(users);
     }
 
     @PutMapping("/users")
     @ResponseStatus(value = HttpStatus.NO_CONTENT)
-    public void updateUser(@RequestBody Users users){
+    public void updateUser(@RequestBody Users users,Principal principal){
         Users dbUsers = usersService.getById(users.getId());
 
-        if (dbUsers !=null)
-            usersService.saveUser(users);
+        boolean proceed = true;
+
+        if(usersService.getById(users.getId()).getRoleId() == rolesService.getByName("ADMIN").getId()){
+            if(!WebSecurityPermissions.isCurrentUser(principal,users.getUserName()))
+                proceed=false;
+        }
+
+        if(WebSecurityPermissions.hasPermission(principal,usersService.getById(users.getId()).getUserName(),"ADMIN")&&proceed){
+            if (dbUsers !=null){
+                users.setPassword(dbUsers.getPassword());
+                if(!WebSecurityPermissions.hasRole("ADMIN")){
+                    users.setRoleId(dbUsers.getRoleId());
+                }
+                usersService.saveUser(users);
+            }
+        }
+        else{
+            throw new NotAuthorizedException("Access Denied, you don't have permissions to access other users data");
+        }
 
     }
 
     @DeleteMapping("/users/{userId}")
-    public int deleteById(@PathVariable int userId){
-        usersService.deleteById(userId);
-        return userId;
-    }
+    public int deleteById(@PathVariable int userId,Principal principal){
 
-    @GetMapping("/users/login")
-    public String loginUser(){
-        return "Welcome!";
+        boolean proceed = true;
+
+        if(usersService.getById(userId).getRoleId() == rolesService.getByName("ADMIN").getId())
+            proceed=false;
+
+        if(WebSecurityPermissions.hasPermission(principal,usersService.getById(userId).getUserName(),"ADMIN")&&proceed)
+        {
+                usersService.deleteById(userId);
+        }
+        else{
+            throw new NotAuthorizedException("Access Denied, you don't have permissions to access other users data");
+        }
+
+        return userId;
     }
 
 }
